+++ contribs/perlapi/libslurmdb/perl/cluster.c
@@ -758,7 +758,7 @@ job_rec_to_hv(slurmdb_job_rec_t* rec, HV* hv)
     STORE_FIELD(hv, rec, qosid,           uint32_t);
     STORE_FIELD(hv, rec, req_cpus,        uint32_t);
     STORE_FIELD(hv, rec, req_gres,        charp);
-    STORE_FIELD(hv, rec, req_mem,         uint32_t);
+    STORE_FIELD(hv, rec, req_mem,         uint64_t);
     STORE_FIELD(hv, rec, requid,          uint32_t);
     STORE_FIELD(hv, rec, resvid,          uint32_t);
     STORE_FIELD(hv, rec, resv_name,       charp);
