+++ src/plugins/select/alps/libemulate/alps_emulate.c
@@ -492,7 +492,8 @@ extern void   free_inv(struct basil_inventory *inv)
 
 extern long basil_reserve(const char *user, const char *batch_id,
 			  uint32_t width, uint32_t depth, uint32_t nppn,
-			  uint32_t mem_mb, uint32_t nppcu, struct nodespec *ns_head,
+			  uint64_t mem_mb, uint32_t nppcu,
+			  struct nodespec *ns_head,
 			  struct basil_accel_param *accel_head)
 {
 	int i;
@@ -501,7 +502,7 @@ extern long basil_reserve(const char *user, const char *batch_id,
 #if _DEBUG
 	struct nodespec *my_node_spec;
 	info("basil_reserve user:%s batch_id:%s width:%u depth:%u nppn:%u "
-	     "mem_mb:%u nppcu:%u",
+	     "mem_mb:%"PRIu64" nppcu:%u",
 	     user, batch_id, width, depth, nppn, mem_mb, nppcu);
 	my_node_spec = ns_head;
 	while (my_node_spec) {
